requirements.txt
GitPython>=3.1.0
javalang>=0.13.0
google-genai>=0.2.0
python-dotenv>=1.0.0